+++ packages/fields-bulma/test/jest.transform.js
@@ -1,3 +1,0 @@
-let babel = require('../babel.config');
-
-module.exports = require('babel-jest').createTransformer(babel);