+++ packages/validator-pwned/js-sha1.d.ts
@@ -1,3 +1,0 @@
-declare module "js-sha1" {
-  export default function(value: string): string;
-}