+++ packages/validator-pwned/CHANGELOG.md
@@ -1,4 +1,4 @@
-# Changelog @react-formilicious/fields-bulma
+# Changelog @react-formilicious/validator-pwned
 
 All notable changes to this library will be documented in this file.
 
@@ -6,4 +6,4 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## 1.0.0
 
-* Created the library using [📚 create-react-prototype](https://github.com/PatrickSachs/create-react-prototype)! @PatrickSachs.
+* Created the library using [📚 create-react-prototype](https://github.com/PatrickSachs/create-react-prototype)! @Patrick Sachs.